Refactor bootloader to allow for kernels > 64k long
Needed quite a refactor and has uncovered the fact that I should really
completely refactor the stage1 loader.

Now uses a long at the start of the kernel image instead of a word, and
only loads 64k at a time to get around a "quirk" of the QEMU BIOS (and
probably many others).

Resolves #17