alloc: Fix bug where memory flags were uninitialised
Memory flags were not set to zero by default, which somehow didn't cause
problems for quite a while! Issues arise when re-using memory, or on
platforms which don't have zeroed memory as an initial state. If you're
unlucky enough to have the MEM_CRITICAL bit of the flags byte in the
block header set, freeing the memory will error out, as it is seen as
critical.