build: Make 'make debug' use i386 QEMU
This is needed because using x86_64 qemu will lead to GDB thinking
(quite rightly) that the architecture is... x86_64. This causes all
sorts of pain. I'm sure there's probably some way to trick GDB into
doing this the right way round, but the simple fix is to use i386 GDB.